Felix HAL: Fixed VibratorTest unit tests errors.
Fixed Felix HAL VibratorHalCs40l26TestSuitePrivate
3 failed tests:
- VibratorTest#on
- VibratorTest#perform/CLICK_LIGHT,
- VibratorTest#perform/HEAVY_CLICK_STRONG.

Moving the mActiveId_mutex to a global variable caused issues with the #perform/HEAVY_CLICK_STRONG test failing to call the HwApi and the #perform/CLICK_LIGHT test failing due to an outdated timeout value. Additionally, the intensifyToVolLevel function was returning an incorrect value, requiring a check to be added to the testing function.

Bug: 373761684
Flag: TEST_ONLY
Test: Ran VibratorHalCs40l26TestSuitePrivate.
Change-Id: Ibcdbc874c5c96644f3f647cc240b5012907253d4
Signed-off-by: Zouberou Sayibou <zouberou@google.com>